# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the Authentication and Authorization Micro-service for the FamTrust Web App.",
        "title": "FamTrust API Backend - Auth",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1/",
    "paths": {
...
        "/images/profile-pic/{imageName}": {
            "get": {
                "description": "Get User Profile Picture",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User-Profiles"
                ],
                "summary": "Get User Profile Picture",
                "operationId": "get-profile-pic",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Picture Filename",
                        "name": "imageName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
        "/profile": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve User Profile Details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User-Profiles"
                ],
                "summary": "Retrieve User Profile Details",
                "operationId": "profile",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.profileSampleResponse200"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.loginSampleResponseError401"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.loginSampleResponseError500"
                        }
                    }
                }
            }
        },
        "/profile/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create New User Profile",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User-Profiles"
                ],
                "summary": "Create New User Profile",
                "operationId": "create-profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User's first name",
                        "name": "firstName",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User's last name",
                        "name": "lastName",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User's biography",
                        "name": "bio",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "User's National Identification Number",
                        "name": "nin",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "User's Bank Verification Number",
                        "name": "bvn",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "User's default group ID",
                        "name": "defaultGroupID",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "User's profile picture",
                        "name": "profilePicture",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
...
        "handlers.loginSampleResponseError401": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Invalid Credentials"
                },
                "status": {
                    "type": "string",
                    "example": "error"
                },
                "statusCode": {
                    "type": "integer",
                    "example": 401
                }
            }
        },
        "handlers.loginSampleResponseError500": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "An error occured"
                },
                "status": {
                    "type": "string",
                    "example": "error"
                },
                "statusCode": {
                    "type": "integer",
                    "example": 500
                }
            }
        },
        "handlers.profileSampleResponse200": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Request successful"
                },
                "profile": {
                    "type": "object",
                    "properties": {
                        "bio": {
                            "type": "string",
                            "example": "The best FamTrust user of all time."
                        },
                        "bvn": {
                            "type": "integer",
                            "example": 35473783473
                        },
                        "createdAt": {
                            "type": "string",
                            "example": "2024-07-22T14:30:00Z"
                        },
                        "firstName": {
                            "type": "string",
                            "example": "Famtrust"
                        },
                        "id": {
                            "type": "string",
                            "example": "a5c9f82e-6b7a-4a53-a81c-82b1e2f453a6"
                        },
                        "lastName": {
                            "type": "string",
                            "example": "Guru"
                        },
                        "nin": {
                            "type": "integer",
                            "example": 35473745433
                        },
                        "profilePictureUrl": {
                            "type": "string",
                            "example": "https://image.famtrust.biz/dkkjieikdjfoej.jpg"
                        },
                        "updatedAt": {
                            "type": "string",
                            "example": "2024-07-22T14:30:00Z"
                        },
                        "userID": {
                            "type": "string",
                            "example": "d38f91b2-dc3b-4f9d-aeb4-7b95c91e9d08"
                        }
                    }
                },
                "status": {
                    "type": "string",
                    "example": "success"
                },
                "statusCode": {
                    "type": "integer",
                    "example": 200
                },
                "token": {
                    "type": "string",
                    "example": "b6d4a7e1d2d841a1afe874a2a5c15d8b"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}